User Control Panel and Extension Form Submission Work
The User Control Panel contains tabs for Tickets, Extensions and editing user account for a JED user. Within the tickets is the ability to view and submit comments on open tickets. Extensions submit is failing due to validate code not being able to correctly identify fields that are behind tabs. Currently investigating.